Move pmin/pmax/epmin/epmax validation from Set to WriteAttributeRequest
